+++ src/Common/RedButton.Common.TeklaStructures/DrawingTable/ListOfDrawingsTable.cs
@@ -1,0 +1,225 @@
+﻿using System;
+using System.Linq;
+using System.Collections.Generic;
+using Tekla.Structures;
+using Tekla.Structures.Geometry3d;
+using Tekla.Structures.Drawing;
+using TSD = Tekla.Structures.Drawing;
+using UI = Tekla.Structures.Drawing.UI;
+
+
+namespace RedButton.Common.TeklaStructures.DrawingTable
+{
+    /// <summary>
+    /// Create List Of drawing table.
+    /// </summary>
+    public class ListOfDrawingsTable
+    {
+        /// <summary>
+        /// Create table of drawings names.
+        /// </summary>
+        /// <param name="drawings">List of drawings</param>
+        /// <param name="headerTitle"></param>
+        public void CreateTable(List<Drawing> drawings,
+            string headerTitle = "Ведомость рабочих чертежей основного комплекта")
+        {
+            var orderedDrawings = drawings.OrderBy(x =>
+            {
+                var number = string.Empty;
+                x.GetUserProperty("ru_list", ref number);
+                double numberDouble = 0;
+                Double.TryParse(number, out numberDouble);
+                return numberDouble;
+            })
+                .ToList();
+                                             
+            var sheetsNumbers = DrawingUtils.SheetNumbers(orderedDrawings);
+            var sheetsNames = DrawingUtils.GetDrawingsTitles(orderedDrawings);
+
+            CreateTable(sheetsNumbers,sheetsNames,headerTitle);
+        }
+
+
+        /// <summary>
+        /// Create table of drawings names.
+        /// </summary>
+        /// <param name="sheetsNumbers">Drawings numbers array</param>
+        /// <param name="sheetsNames">Drawings names array</param>
+        /// <param name="headerTitle">Table title</param>
+        public void CreateTable(double[] sheetsNumbers, string[] sheetsNames,
+            string headerTitle = "Ведомость рабочих чертежей основного комплекта")
+        {
+            //======Table settings==========
+
+            //Header
+            var headerHeight = 15;
+            string[] headerText = { "Лист", "Наименование", "Примечание" };
+
+            //Row
+            var rows = sheetsNames.Length;
+            var rowHeight = 8;
+
+            //Columns width
+            double[] columnWidth = { 15, 140, 30 };
+            var tableWidth = columnWidth.Sum();
+
+            //=====Text settings===========
+            var textHeight = 2.5;
+            var textFont = "Arial Narrow";
+
+
+            //=====Insert table=======
+            try
+            {
+                DrawingHandler drawingHandler = new DrawingHandler();
+                if (drawingHandler.GetConnectionStatus())
+                {
+                    UI.Picker picker = drawingHandler.GetPicker();
+                    Point startPoint = null;
+                    ViewBase viewBase = null;
+
+
+                    picker.PickPoint("Укажите точку вставки таблицы", out startPoint, out viewBase);
+
+                    DrawTableLines(viewBase, startPoint, tableWidth, rows, rowHeight, headerHeight, columnWidth);
+
+
+                    DrawTableText(sheetsNumbers, sheetsNames, textHeight, textFont, viewBase, startPoint, tableWidth, rowHeight, headerTitle, headerText, columnWidth, headerHeight);
+
+
+                    drawingHandler.GetActiveDrawing().CommitChanges();
+                }
+
+            }
+            catch (Exception exp)
+            {
+                // ignored
+            }
+        }
+
+
+        private void DrawTableLines(ViewBase viewBase, Point startPoint, double tableWidth, int rows, int rowHeight,
+            int headerHeight, double[] columnWidth)
+        {
+            //=========Drawing table=========
+
+            TSD.Line.LineAttributes lineAtt = new TSD.Line.LineAttributes();
+            LineTypeAttributes lineTypeAtt = new LineTypeAttributes(LineTypes.SolidLine, DrawingColors.Black);
+            lineAtt.Line = lineTypeAtt;
+
+            //Table Header
+            TSD.Line headerLine = new TSD.Line(viewBase, startPoint,
+                new Point(startPoint.X + tableWidth, startPoint.Y, startPoint.Z), lineAtt);
+            headerLine.Insert();
+
+
+            //Horizontal lines
+            for (int i = 0; i < rows + 1; i++)
+            {
+                TSD.Line upperLine = new TSD.Line(viewBase,
+                    new Point(startPoint.X, startPoint.Y - (i * rowHeight + headerHeight), startPoint.Z),
+                    new Point(startPoint.X + tableWidth, startPoint.Y - (i * rowHeight + headerHeight), startPoint.Z), lineAtt);
+                upperLine.Insert();
+            }
+
+
+            //Vertical lines
+            TSD.Line firstVerticalLine = new TSD.Line(viewBase, startPoint,
+                new Point(startPoint.X, startPoint.Y - (rows * rowHeight + headerHeight), startPoint.Z), lineAtt);
+            firstVerticalLine.Insert();
+
+            double currentWidth = 0;
+            for (int i = 0; i < columnWidth.Length; i++)
+            {
+                currentWidth += columnWidth[i];
+
+                TSD.Line verticalLine = new TSD.Line(viewBase,
+                    new Point(startPoint.X + currentWidth, startPoint.Y, startPoint.Z),
+                    new Point(startPoint.X + currentWidth, startPoint.Y - (rows * rowHeight + headerHeight), startPoint.Z),
+                    lineAtt);
+                verticalLine.Insert();
+            }
+        }
+
+
+        private void DrawTableText(double[] sheetNumbers, string[] sheetNames, double textHeight, string textFont,
+            ViewBase viewBase, Point startPoint, double tableWidth, int rowHeight, string headerTitle, string[] headerText,
+            double[] columnWidth, int headerHeight)
+        {
+            //=========Drawing text=========
+
+            Text.TextAttributes textAtt = new Text.TextAttributes();
+            textAtt.Alignment = TextAlignment.Left;
+            textAtt.Font = new FontAttributes(DrawingColors.Black, textHeight, textFont, false, false);
+
+            //Header text
+
+            Text.TextAttributes titleAtt = new Text.TextAttributes();
+            titleAtt.Alignment = TextAlignment.Left;
+            titleAtt.Font = new FontAttributes(DrawingColors.Black, textHeight + 2, textFont, false, false);
+            TSD.Text titleWord = new TSD.Text(viewBase,
+                new Point(startPoint.X + (tableWidth / 2), startPoint.Y + rowHeight, startPoint.Z), headerTitle, titleAtt);
+            ;
+            titleWord.Insert();
+
+
+            TSD.Text headerWord = null;
+            double insertXcoord = 0;
+
+            for (int i = 0; i < headerText.Length; i++)
+            {
+                if (i > 0)
+                {
+                    insertXcoord += columnWidth[i - 1];
+                }
+
+                headerWord = new TSD.Text(viewBase,
+                    new Point(startPoint.X + insertXcoord + (columnWidth[i] / 2), startPoint.Y - (headerHeight / 2),
+                        startPoint.Z), headerText[i], textAtt);
+                headerWord.Insert();
+            }
+
+
+            //Column number
+            TSD.Text sheetNumber = null;
+
+            for (int i = 0; i < sheetNumbers.Length; i++)
+            {
+                sheetNumber = new TSD.Text(viewBase,
+                    new Point(startPoint.X, startPoint.Y - (i * rowHeight) - headerHeight, startPoint.Z),
+                    sheetNumbers[i].ToString(), textAtt);
+                sheetNumber.Insert();
+
+                sheetNumber.MoveObjectRelative(new Vector(new Point((columnWidth[0] / 2), (-rowHeight / 2),
+                    0))); //TextMoveVector(sheetNumber, rowHeight));
+                sheetNumber.Modify();
+            }
+
+            // SheetNameColumn
+            TSD.Text sheetText = null;
+
+            for (int i = 0; i < sheetNames.Length; i++)
+            {
+                if (!sheetNames[i].Equals(""))
+                {
+                    sheetText = new TSD.Text(viewBase,
+                        new Point(startPoint.X + columnWidth[0], startPoint.Y - (i * rowHeight) - headerHeight, startPoint.Z),
+                        sheetNames[i], textAtt);
+                    sheetText.Insert();
+
+                    sheetText.MoveObjectRelative(TextMoveVector(sheetText, rowHeight));
+                    sheetText.Modify();
+                }
+            }
+        }
+
+
+        private Vector TextMoveVector(TSD.Text text, double rowHeight)
+        {
+            Point boundLowerLeft = text.GetObjectAlignedBoundingBox().LowerLeft;
+            Point textIstPoint = text.InsertionPoint;
+
+            return new Vector(new Point(textIstPoint.X - boundLowerLeft.X, -rowHeight / 2, 0));
+        }
+    }
+}
